Create BaseAgent, organize files
